feat: 봉투 생성 응답값 변경

+++ data/src/main/java/com/susu/data/data/repository/FriendRepositoryImpl.kt
@@ -13,7 +13,7 @@ class FriendRepositoryImpl @Inject constructor(
     override suspend fun createFriend(
         name: String,
         phoneNumber: String?,
-        relationshipId: Int,
+        relationshipId: Long,
         customRelation: String?,
     ): Long = friendService.createFriend(
         FriendRequest(
